2/9/2014
Its been awhile. After finishing the Missing Manual PHP & MySQL book (I
highly recommend it, although it uses mysql instead of mysqli) I can
start to construct a user account system which all the bells and
whistles. The 'project' has been renamed to Celestial Wars, with the
main additions of the homepage and registration page written up, but not
yet functioning. I am planning to spend the next few days exploring the
usefulness that is AJAX.